+++ application/src/main/kotlin/application/backwardCompatibility/BCCExampleValidationStatus.kt
@@ -1,0 +1,9 @@
+package application.backwardCompatibility
+
+internal data class BCCExampleValidationStatus(
+    val areExamplesInvalid: Boolean,
+    val hasUnloadableExamples: Boolean
+) {
+    val hasErrors: Boolean
+        get() = areExamplesInvalid || hasUnloadableExamples
+}